+++ react-app/src/components/MessageFeed/index.js
@@ -1,0 +1,3 @@
+import MessageFeed from "./MessageFeed";
+
+export default MessageFeed;
